move images slider to widgets
