+++ lib/views/views.dart
@@ -1,5 +1,4 @@
 export 'package:pokemon_mobile_course/views/pokemon_list.dart';
 export 'package:pokemon_mobile_course/views/splash_screen.dart';
-export 'package:pokemon_mobile_course/views/pokemon_images_slider.dart';
 export 'package:pokemon_mobile_course/views/pokemon_details.dart';
 export 'package:pokemon_mobile_course/views/my_pokemons_view.dart';
