+++ lib/widgets/pokemon_images_slider.dart
